% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE-ERROR-package.R, R/ATE.ERROR-package.R
\docType{package}
\name{ATE.ERROR-package}
\alias{ATE.ERROR}
\alias{ATE.ERROR-package}
\title{ATE.ERROR: Estimating ATE with Misclassified Outcomes and Mismeasured Covariates}
\description{
Addressing measurement error in covariates and misclassification in binary outcome variables within causal inference, the 'ATE.ERROR' package implements inverse probability weighted estimation methods proposed by Shu and Yi (2017, \doi{10.1177/0962280217743777}; 2019, \doi{10.1002/sim.8073}). These methods correct errors to accurately estimate average treatment effects (ATE). The package includes two main functions: ATE.ERROR.Y() for handling misclassification in the outcome variable and ATE.ERROR.XY() for correcting both outcome misclassification and covariate measurement error. It employs logistic regression for treatment assignment and uses bootstrap sampling to calculate standard errors and confidence intervals, with simulated datasets provided for practical demonstration.

Addressing measurement error in covariates and misclassification in binary outcome variables within causal inference, the 'ATE.ERROR' package implements inverse probability weighted estimation methods proposed by Shu and Yi (2017, \doi{10.1177/0962280217743777}; 2019, \doi{10.1002/sim.8073}). These methods correct errors to accurately estimate average treatment effects (ATE). The package includes two main functions: ATE.ERROR.Y() for handling misclassification in the outcome variable and ATE.ERROR.XY() for correcting both outcome misclassification and covariate measurement error. It employs logistic regression for treatment assignment and uses bootstrap sampling to calculate standard errors and confidence intervals, with simulated datasets provided for practical demonstration.
}
\author{
\strong{Maintainer}: Aryan Rezanezhad \email{Aryan.rzn@gmail.com}

Authors:
\itemize{
  \item Grace Y. Yi \email{gyi5@uwo.ca}
}

}
\keyword{internal}
